requirements.txt
ccxt
pandas
pymongo
python-dateutil
requests
colorama
